Bug 137299: POST form data is missing with a custom NSURLProtocol
Allow to send HTTP body form data over IPC in StartLoading message.

Add a parameter to ResourceRequest IPC argument encoder,
which controls if the NSURLRequest.HTTPBody is stripped.
When sending a CustomProtocolManagerProxy::StartLoading message,
encode message parameters so that the body is not stripped.
In other cases the HTTP body is stripped (for performance reasons).

Add test case WebKit2CustomProtocolsTest.FormPostBodyExists.

In addition:
Refactor ArgumentCoder<ResourceRequest>::encodePlatformData
in order to remove duplicate code.